+++ TP_MVC/TP/Controllers/AnimalController.cs
@@ -195,6 +195,11 @@ namespace TP.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            var animalExiste = _context.Adopcion.FromSqlRaw("select * from adopcion where idAnimal = " + id).ToList().Count;
+            if (animalExiste > 0)
+            {
+                return NotFound($"Error: El animalito tiene una adopción registrada. Elimine la adopción para poder eliminar el animalito.");
+            }
             var animal = await _context.Animal.FindAsync(id);
             _context.Animal.Remove(animal);
             await _context.SaveChangesAsync();
